Add base_url parameter.
The base_url is the url the user is expected to use for the website. It
is also the base of the address that will be passed to the OAUTH
providers (Google, GitHub, and in the future(?) Facebook). The URL must
match the allowed redirect URIs in the project's OAUTH registration.

Also, by explicitly specifying the URL to use, it should also help
prevent the problem of people trying to use the 0.0.0.0 address for
browsers (like Chrome) that don't allow it.

Fixes PythonIreland/broadgauge#24